requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
psycopg[binary]
requests
